Add helper script for setting up test VM
This adds a script to help make it as fast as possible to get a new test
vm up and running. It only supports Ubuntu so far, but the hope is to
expand it to cover other deployment scenarios.

This is probably only useful as we work on getting things ready. We will
hopefully have a better deployment story once we get closer to full
release.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>