Update getting started guide to drop tkg cli
We should now be able to do everything with the tanzu CLI. This updates
installation instructions to move over to use the tanzu commands.

Also updates the download instructions with the expected version of the
next release. This may need to be updated, but I thought it would be
better to have invalid links than to have someone download the old
version that would not work with the current instructions.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>