Clean up release dependency builds
This moves the build of the core repo and other dependencies outside of
the repo root. Having it local causes problems since it doesn't get
fully cleaned up. Running things like `make check` after building a
release then results in a lot of false errors due to the extra files
present in the repo.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>