Update build scripts to exit on failure
Some scripts were calling return on failure, which would then error due
to not being in a function from which to return. The scripts are
currently set to continue, so unless you saw it in the logs, you
wouldn't know there was an issue.

This updates the scripts to `exit` rather than `return` when not running
in a function.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>