Increase `make lint` timeout to 5 minutes
The `make lint` target calls golangci-lint to perform various linting.
On development machines, resources are usually pretty good and this
linting is performed fairly fast. But in the GitHub Actions there are a
lot less resources available and things take a bit longer to run.

The default timeout for golangci-lint is 1 minute. This is overridden in
our Check job configuration, so when run on normal PRs, this check has
plenty of time to run.

We run into issues with the job that is run on new commits being pushed
to main. This job configuration wants to run all of our checks, so
rather than using the golangci-lint action, it just calls `make check`,
which pulls in the lint target and other checks.

Because we had been calling golangci-lint there without passing in a
--timeout, it falls back to the 1 minute default timeout and ends up
failing when run in this Action.

This increases the timeout in the Makefile to 5 minutes to avoid hitting
this on merge commits.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>