Redirect sed output instead of copy/edit
This updates the addon creation script to redirect the output of using
sed on the README template to the new addon's directory, rather than
copying the template over, then editing in place. This helps avoid one
of the cases where in-place editing with sed can cause issued between
different platforms.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>