Better handle sed differences between macoOS and Linux (#419)
* Better handle sed differences between macoOS and Linux

The sed version on Mac is different from Linux. There is no clean way to
work on both avoiding creating a backup file, so rather than fight it,
just create a backup file and remove it right away.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>

* Redirect sed output instead of copy/edit

This updates the addon creation script to redirect the output of using
sed on the README template to the new addon's directory, rather than
copying the template over, then editing in place. This helps avoid one
of the cases where in-place editing with sed can cause issued between
different platforms.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>