Use common table output interface
We have a few commands that need to write out a table of their results.
There are also several in the core repo that may or may not end up in
this repo too. Rather than leaving it to each command, and command
implementer, to figure out and format their data in each location, this
adds a common simple interface to use to generate and output the data in
a consistent table format.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>