Add mutate function for controllerutil.CreateOrUpdate
These calls require a MutateFn to be passed in. This is normally used to
mutate the API object, but we do not have a need to do that right now.

In case we do have a need in the future, and to have something to make
the controllerutil calls happy, this defines a noop function that can be
used for these instances.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>