Raise rangeValueCopy limit for gocritic linting
gocritic has good checks to make sure the code is not copying over large
amounts of memory, but the default threshold for it is fairly low. We
have a few for loops that iterate through items that really are not much
of an impact. Rather than restructuring code to get around this check,
this bumps the threshold to only alert on larger copies.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>