Use common table output interface (#440)
* Use common table output interface

We have a few commands that need to write out a table of their results.
There are also several in the core repo that may or may not end up in
this repo too. Rather than leaving it to each command, and command
implementer, to figure out and format their data in each location, this
adds a common simple interface to use to generate and output the data in
a consistent table format.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>

* Raise rangeValueCopy limit for gocritic linting

gocritic has good checks to make sure the code is not copying over large
amounts of memory, but the default threshold for it is fairly low. We
have a few for loops that iterate through items that really are not much
of an impact. Rather than restructuring code to get around this check,
this bumps the threshold to only alert on larger copies.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>

* Update kapp-controller and dependencies

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>