Only run Check Build job on main
Right now, the main branch is the only one where we need to perform a
full build on PRs to validate. There is a lot of work happening in the
docs branch. These doc-only changes are not touching code, so waiting on
builds to complete is unnecessary and consuming Action time.

We may have something like release branches in the future, in which case
we can add them or change how this is handled.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>