Remove old docs to avoid confusion (#576)
The current docs are being developed under the docs branch. This branch
will eventually be merged back in to main.

There are various updates to existing design docs in that branch. To
avoid confusion and missed updates from those that are not aware of
this, this removes the old docs and updates references. All of these
docs will be present under docs/site, so this should not cause conflicts
when it comes time to merge back.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>