GitHub workflow to copy package readmes to docs #881
Adds a GitHub workflow that triggers on changes to package readme files.

If the package is in the user-managed repo, the readme file will be copied
to the docs directory. The documentation table of contents will be
regenerated to capture package deletions/addtions. Finally the workflow
will create and merge a PR for the documentation updates.

Signed-off-by: Nicholas Seemiller <nseemiller@vmware.com>