Only list directories for prep-build-cli
In our prep-build-cli make target we iterate through all CLI plugins to
run `go mod download` ahead of actually building to make sure all
dependencies have been fetched. This had been iterating through
everything in that directory, but a README file was added that would
cause error output when the script tried to cd into it as a directory.

This makes a minor update to the command to only iterate through the
directories in the plugin directory instead of all list results.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>