[e2e][aws] Delete local kind bootstrap cluster responsible for standalone cluster creation, before standalone cluster deletion (#1270)
During the TCE standalone cluster deletion process, the local kind bootstrap
cluster that created the standalone cluster could be up and running for some
reason, for example due of some issue. This local kind bootstrap cluster with
all the cluster API providers could be repeatedly reconciling to keep the
standalone cluster resources up and running to keep the standalone cluster up
and running. This can hinder the standalone cluster deletion process. To avoid
this situation, we would have to delete the local kind bootstrap responsible
for standalone cluster creation

Signed-off-by: Karuppiah Natarajan <karuppiahn@vmware.com>