Remove unnecessary Go setup steps in gh actions workflows (#1389)
* [ci] Remove unnecessary Go setup step from github actions workflow

Only the carvel tools are needed for update packages workflow, which can
be installed using prebuilt binaries or installed using
GitHub Actions available in the GitHub Actions marketplace
and we dont need Golang to be installed to install those
unless we want to build from source and install them using
`go get`

check-shell uses shellcheck tool and we use the installed
binary or use Docker image to invoke it. We don't need
Golang setup for it

check-mdlint uses mdlint tool and we use mdlint Docker image
to invoke it. We don't need Golang setup for it.

check-lint workflow uses golangci-lint tool and we use golangci-lint
GitHub Action to invoke it. We don't need Golang setup for it.