Fix linting coverage and issues
The linting job runs within the scope of a module, so for each module
added to the repo, the linting target would need to be updated to cover
each one. Modules have been added, but linting wasn't covering them
because this was missed.

This updates the linting target to pick up all submodules in the repo.
It also fixes uncovered by this additional coverage.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>