fix two erroneous directory aliased
Windows uninstall left behind the .cache and .config directories because an incorrect LocalAppData directory alias was used - changed to use correct USERPROFILE alias.
Also prevented outputing <Uninstall complete> if an error occurs and now return an exit code to allow the script to be potentially be used in test automation. Also output an elapsed time of script execution.

Fixes #1475

Signed-off-by: Steve Wong <wongsteven@vmware.com>