add custom wsl kernel docs
This commit adds the build steps for creating a custom kernel to allows
for local (CAPD-based) installations to work on Windows. This custom
kernel is required to support antrea's usage of the
CONFIG_NETFILTER_XT_TARGET_CT configuration in the kernel.

This custom kernel will not be required in the long term. The relevant
antrea issue is https://github.com/antrea-io/antrea/issues/2635.

Signed-off-by: joshrosso <rossoj@vmware.com>