Insert go module version of Sonobuoy at build time
We cannot associate a Go module version with a commit; even if we use
commits when pulling, if a commit is referenced by a git tag, Go will
choose the git tag as the version.

Therefore, we'll just query Go for the version it knows about from the
modules, using a version or pseudoversion depending on which is there.
Pseudoversions carry SHA information, should we need it.

Signed-off-by: Nolan Brubaker <brubakern@vmware.com>