Hide header link symbols by default
We currently show the paragraph symbol on all headers to indicate they
are directly linkable. This is great, but can add a lot of visual
clutter to the screen when reading the docs.

Following what I've seen work well for other docs sites, this changes
those links to be "hidden" by being set to the same color as the
background until hovering over the title. On hover, the symbol becomes
visible as a link.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>