fix inconsistent font size in docs code blocks
This commit ensures the code blocks set the font size. With out it, font
sizes were inherited by parent CSS classes. This made code blocks that
were indented under a step (ie <ul> or <ol>) to take on that style and
have a larger font that other code blocks on the page.

Signed-off-by: joshrosso <rossoj@vmware.com>