Enable linting or docs/site markdown files
We had mdlinting disabled for the hugo site while we worked on the
content. Now that we are close to stabilizing it, we should enable the
linting to make sure our content is well formed and syntactically
correct.

This enabled mdlint on more directories and addresses the many issues
that has found. Some exclusions were kept for problems files such as the
fractional "asset" content until we can figure out how to handle those,
but issues have been fixed in them as much as possible.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>