Add links to mdlint rules documentation
When errors are encountered in the mdlint jobs, if you aren't very
familiar with markdown it can be confusing understanding what the error
is and how you are supposed to fix it. This adds a couple breadcrumbs to
the make target and mdlinting script to point to documentation that goes
over what each error is, with examples of formatting that can cause them
and suggestions on how to fix them.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>