add automation for azure management and workload cluster e2e test (#1833)
- use azure utils script in azure standalone cluster automation script to reuse functions
- check required environment variables in util functions too. This helps one to use util
functions standalone in any place. The utils functions only check the presence of
environment variables used in that util function, but not
the ones used in it's dependencies which are also util functions that contain their own
environment variables check. This helps avoid duplicating the check but also preserves the
idea of failing fast to a good extent when pre-requisites aren't met. The top level standalone
and management + workload cluster scripts check the presence of
environment variables used in their dependencies though, which are util functions.
This is to provide a fail fast strategy as the top level scripts are really long proceses and
we would like to fail really really fast as the first thing if something is missing

Signed-off-by: Karuppiah Natarajan <karuppiahn@vmware.com>