Standardize LICENSE file
The LICENSE file should only contain the Apache license, with nothing
extra. Automated checkers will recognize the plain license, but having
it modified will trigger warnings that often require manual review.
Since we are not modifying the use of the license in any way, we should
only have the license text itself in our LICENSE file.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>