Wire in config struct for local cluster create
This starts the process for adding the use of a config struct for
managing all of the local cluster configuration settings. For now it
just hardwires it in to the creation call, but will be expanded to be
something passed in with different settings that can then be used to
change how we create the cluster.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>