Handle empty CLI default config values
Initializing the local cluster configuration was checking for the
presence of config settings, but not whether they actually contained a
value. This is a problem with the real CLI processing since the default
values set on all options ends up being "", resulting in default values
being used unless passed in as a command line argument.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>