Add port mapping to sa cluster creation
This adds the ability to provide a list of ports to map from container
to host when creating a new standalone cluster. Multiple port mappings
can be provided by specifying the --port-map or -p switch and giving it
in the Docker format of container:host/protocol.

Note: the -p switch had previously been used as the short form for
--provider. This removes that and makes -p the short form for --port-map
since that is more consistent with how tools like the docker CLI handle
mapping or publishing ports.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>