Update get-deps to use "go mod tidy"
Most get-deps targets were using "go mod download", which isn't as
accurate as "go mod tidy" for evaluating dependencies. From one of the
go maintainers:

```
go mod download is downloading all of the modules in the dependency
graph, which it can determine from reading only the go.mod files. It
doesn't know which of those modules are actually needed to satisfy a
build, so it doesn't pull in the checksums for those modules (because
they may not be relevant).

On the other hand, go mod tidy has to walk the package graph in order
to ensure that all imports are satisfied. So it knows, for a fact,
without doing any extra work, that the modules it walks through are
needed for a go build in some configuration.
```

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>
Signed-off-by: Karuppiah Natarajan <karuppiahn@vmware.com>