Windows uninstall: remove tanzu from PATH
In our install.bat file at the end of installation we prompt the user to
add the Tanzu CLI directory to their system path. During uninstall we
don't mention anything.

This updates the uninstall.bat file to emit a note about modifying the
PATH to remove this if uninstall completes successfully.

While updating, changed the line endings to CRLF since that is the
native, expected line endings for Windows.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>