Enable markdownlinting for README (#2509)
Linting was disabled for the entire file since we want to start with
html for our logo image. But we really only need to disable that one
check.

This makes the disable more specific to get by the raw html inclusion,
enabling checks for the rest of the file. A few minor lint issues are
addressed.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>