Bumps vulnerable dependencies for dependabot alerts (#2625)
* Bumps vulnerable dependencies for dependabot alerts

- Ignores go.sum files in misspell checker

Signed-off-by: John McBride <jmcbride@vmware.com>

* Update golangci-lint to 1.43.0

The version of golangci-lint we were using included a linter that had a
panic issue when running under go 1.17. This updates the version of
golangci-lint we use to a newer release that pulls in a set of updated
linters that fixes this, as well as adding improved lint checks.

One notable advantage of this is there is now an error emitted if the
deprecated ioutil package is used.

This update also uncovered a few other issues. Fixes for those are
included.

Go.sum files were out of date for some modules. We don't check that
after running out lint jobs, but running locally resulted in modified
files. So this picks up those changes too.

One thing of note is this also drops the gomnd lint check. Magic number
checks are useful, but in the majority of instances that this linter
pointed out, we just nolint the instance and ignore it. Rather than
enabling it globally and just ignoring it everywhere, this just disables
the global use of it.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>
Signed-off-by: John McBride <jmcbride@vmware.com>

Co-authored-by: Sean McGinnis <smcginnis@vmware.com>