add antrea configuration value for serviceCIDRv6
This change is being made to support dualstack configurations.
When running single stack, v4 or v6, only the serviceCIDR needs to be
set.

When running dualstack, both a v4 and v6 CIDR needs to be set in the
ConfigMap antrea-agent.conf serviceCIDR and serviceCIDRv6 config variables.
This change makes that possible.

An example of how the config vars are set can be found in the framework
repo:
https://github.com/vmware-tanzu/tanzu-framework/blob/5338922f196a8cbb33114df45bf9f725c5f9579c/pkg/v1/providers/ytt/02_addons/cni/antrea/antrea_addon_data.lib.yaml#L41-L47

Signed-off-by: Tyler Schultz <tschultz@vmware.com>
Co-authored-by: Aidan Obley <aobley@vmware.com>