Add TCE proposal for creating community-edition releases
At the time of writing, Tanzu Community Edition (TCE) relies on the Bill
of Materials (BOMs) produced for VMware's paid product, Tanzu Kubernetes
Grid (TKG). This means that creation of clusters, both management and
workload, **always** use the same packages and versions as the rest of
TKG. This binding to a TKG BOM was a decision made to ensure we could
release TCE. Now we are at stage where we must unbind from TKG's BOM and
begin producing TCE-specific BOMs.

Signed-off-by: joshrosso <rossoj@vmware.com>