unmanaged-cluster: remove usage of grey log text
This commit removes the usage of grey log text in favor of no text
stylization in bootstrapping.

This approach introduces ColorNone rather than trying to stylize under
a white value. This is because many modern terminal emulators let you
overwrite the meaning of white. This will result in the terminal's
default "no style". ColorBrightBlack has been removed since it is no
longer used; it can be re-introduces should it be used in the future.

This usage of they grey text was intended to focus the eye on
Event/Eventf commands. However, based on terminal stylization, we found
that this darker text had a high-variance in visibility. Atop that, we
should do more research in the realm of accessibility before dimming out
such text as a default value.

Signed-off-by: joshrosso <rossoj@vmware.com>