Address lint issues (#2829)
With the merge into this development branch from main, we've picked up
the latest golangci-lint settings. These point out a few issues that
were not found with the older version of the linters. We've also been
ignoring a few things, but since we've switched from golint to revive
for some checks they were being highlighted again.

To address new issues and to stop ignoring others, this makes various
minor updates to avoid linting complaints.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>