Fork output writer to avoid bold table headers (#2841)
We had been using the OutputWriter from tanzu-framework. In the release
we are using, the header values for table columns in that version would
bold the text, resulting in garbage characters when doing things like
`watch` to wait for conditions in the output. This has since been fixed,
but is not available in a tanzu-framework release we can use yet.

Until we are able to bump up to a version of tanzu-framework that has
the fix, this temporarily duplicates that code locally. This can be
removed in a future release.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>