Change tty control flag to --tty-disable (#2850)
Bool flags for cobra commands are defaulted to true when present, and
the syntax of needing an = to set the value to something other than that
isn't the most intuitive.

Rather than have our tty control flag default to true, requiring the
user to explicitly set the value to switch it to false, this simplifies
things by changing the flag to `--tty-disable`. That makes its usage a
little simpler in that when it is actually used, the intent of using it
is to disable the TTY formatting in our output. So now it can just be
used as a simple flag.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>