Don't error if unable to read unmanaged clusters
If we attempt to list unmanaged clusters before any have been created,
the config directory does not exist yet and the command errors out
stating that the directory does not exist.

Rather than erroring, we can safely assume that if the directory can't
be read, there just aren't any clusters to list. This updates the list
logic to return an empty list in this case rather than an error.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>