unmanaged-cluster: support no CNI and failed CNI (#2941)
This commit adds support for creating clusters without installing CNI. It
also enables cluster creation to complete successfully when a CNI
package fails to resolve.

For creating clusters without CNI, a user may set the CNI configuration
value to "none". For example, cluster creation can occur as follows:

tanzu uc create testme --cni=none

Users will receive a warning message that no CNI was installed, and the
installation will complete.

For when a CNI is specified but is invalid, for example:

tanzu uc create testme --cni=beepboopnotreal

Users will receive a warning message that no CNI package was found, but
the install will still complete. This is because the cluster is still
successfully started and the user could choose to remedy or troubleshoot
after bootstrapping.

Signed-off-by: joshrosso <rossoj@vmware.com>