Use regex for copyright header year match (#2764)
Our policy is to have the copyright header reflect the year the file was
created, or a range from when it was created to the most recent
modification. With the existing header linting config in place, it would
expect that range to be updated even if there were no code changes,
which is not what we want to enforce.

This changes the linting config to use a regex that will accept either a
range or an individual year. This also matches what has been done in the
tanzu-framework repo.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>