Fix goheader year regex
A space was added after the single year regex so it wouldn't match
something like "2021-9000", but it was not accounted for in the
resulting template.

This updates the regex to capture that space after the single year
(2022) or multi-year (2021-2022) values and updates the template so it
does not end up expecting two spaces after the year match.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>