Enable misspell checks on docs
We had been skipping the docs source files in our `misspell` checks,
presumably because the site had a lot of active churn as we were
preparing for launch. Now that the site has stabilized, it would be good
to run these checks to catch spelling errors in our documentation.

This removes the docs tree from the list of ignored paths for the check
and addresses the spelling errors that uncovered.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>