Correct casing of /etc/kubernetes path (#3206)
Our troubleshooting docs walk through how to exec into a pod in order to
run kubectl commands. As part of this, we have the user copy the
admin.conf file over so it doesn't need to be specified as the
`--kubeconfig` when running commands. The example command we give for
them to copy and paste had a type where "Kuberenetes" was capitalized in
that path, so just pasting that command would fail to find the file.
This changes that path to the correct all-lowercase
`/etc/kubernetes/admin.conf`.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>
(cherry picked from commit 21991417e2978f07183d2811e2fb299d46b1fd83)