Merge "next" post-v0.10.0 development to main (#3089)
* Move plugin definition locally (#3021)

This moves the PluginDescriptor and some helper functions to be light
weight, locally defined interfaces with the community-edition repo.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>

* Update next branch with TF 0.16.0 (#3149)

Co-authored-by: github-actions <github-actions@github.com>

* Unmanaged-cluster: resolve kbld images (#3133)

- Using kbld libs, resolves the `image:` tags correctly from the
  `.imgpkg/images.yaml` files found in the kapp bundle
- `Image` struct now has MergedManifest instead of ResolvedBytes.
  MergedManifest is a single slice of bytes that represents the resolved
  complete resolved, merged ytt / kbld manifest

Signed-off-by: John McBride <jmcbride@vmware.com>

* Fix update GCP buckets (#3162)

Co-authored-by: github-actions <github-actions@github.com>

* Revert publish for normal builds, remove distro on gcp bucket (#3164)

Co-authored-by: github-actions <github-actions@github.com>

* Removes standalone-cluster plugin and all associated bits (#3007)

- Removes standalone-cluster plugin
- Removes standalone-cluster e2e test for all providers
- Removes standalone-cluster release and install

Signed-off-by: John McBride <jmcbride@vmware.com>

* Removal of standalone support mode in diagnostic plugin (#2922)

Signed-off-by: Vladimir Vivien <vivienv@vmware.com>

* remove standalone from plugin build (#3169)

Signed-off-by: joshrosso <rossoj@vmware.com>

* Only build platforms needed in action checks (#3110)

We have several "check" jobs that run on new PRs. These jobs are
currently running `make release` to get the binary to use to perform
those checks.

The GitHub Actions run on a Linux AMD64 platform, but by default the
`make release` target builds for all supported platforms. We end up
waiting a long time while these unnecessary platform targets are built,
to then just pull out the linux-amd64 binary to perform our tests.

This updates the `make release` targets called in these actions to only
build the linux-amd64 binary that we need, saving some time by not
waiting on platform targets we just throw away.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>

* Document chocolatey release commands (#3175)

Signed-off-by: Nolan Brubaker <brubakern@vmware.com>

* Bump go version to 1.17 (#2865)

The tanzu-framework repo has bumped its version of go to 1.17 [0].
This follows suite and updates the go version in all go.mod and
github workflows to specify go 1.17.

[0] https://github.com/vmware-tanzu/tanzu-framework/commit/b2816e1ae30f3d8cf0c201eda05fcd89ea6c5924

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>

* Handle go 1.16 and 1.17 deps for unmanaged plugin (#3218)

This updates the `go mod tidy` calls for the unmanaged cluster plugin to
be compatible with both 1.16 and 1.17. We require 1.17, but the builder
tries to do a `go mod tidy` and will fail without any flags provided due
to some transitive dependencies.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>

* Use stg repo for v0.16.0 (#3227)

Until images get promoted to the production harbor

Signed-off-by: John McBride <jmcbride@vmware.com>

* Upgrade Tanzu-Framework to v0.17.0 (#3231)

Signed-off-by: John McBride <jmcbride@vmware.com>

Co-authored-by: David vonThenen <12752197+dvonthenen@users.noreply.github.com>
Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: John McBride <jmcbride@vmware.com>
Co-authored-by: Vladimir Vivien <vivienv@vmware.com>
Co-authored-by: joshrosso <rossoj@vmware.com>
Co-authored-by: Nolan Brubaker <brubakern@vmware.com>