Fail install if sudo access fails
If someone installing TCE either does not have sudo access, or enters
the wrong password, it is currently ignored and the script continues on
but does not actually complete the install.

This adds a check on the sudo attempt to make sure it is successful. If
not, it gives an error and exits will a failure return code.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>