Reenable submodule linting
During restructuring to add a root go.mod, some debugging was left in
the Makefile definition for the "lint" target, leaving anything but the
root linting off by having it just echo the command.

This reenables linting for those submodules and addresses some of the
things that have crept in since then.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>