Fix install.sh sudo failure handling
Error handling was added to exit if we are unable to sudo when trying to
install the tanzu binary into a system path. This had an error though
where the "exit 1" would be executed outside of the "or", so whether
sudo worked or not it would still exit with a failure.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>