Windows: hide error deleting non-existent path
Our install.bat script tries to clean up a cache directory during
install to make sure there aren't any stale cached entries. This works,
but in cases where that directory does not exist, the command still
would emit the error message:

```
The system cannot find the path specified.
```

This may cause confusion to end users thinking something has gone wrong,
and it does not matter because ultimately we do not want the path
specified to exist anyway. So this change swallows any stderr output so
this result is silently ignored.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>