Fix shellcheck shell script linting
We moved the location of the `check-shellsheck.sh` script to be another
level deep in the folder structure. There is a line in the script that
mounts one level up to the container running the shell linting, so with
this new structure it would no longer find all of our script files.

This updates the line to reflect the new location in the repo and
addresses the issues that have been introduced since the move.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>