Update path for /usr/bin/env reference (#3501)
Install scripts were updated to use the preferred shebang of referencing
`bash` through the use of `env`. The location of `env` is somewhat
standardized as `/usr/bin/env`, but the update referred to `/bin/env`.

This is fine on most systems, but at least our GitHub Action runners
aren't able to resolve and run the install script properly with the
current path. This updates those shebangs to point to the canonical
locaiton of `env`.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>