Rename cli plugin publish jobs to support macOS
The version of `make` on macOS does not support deep pattern matching.
This was causing problems with having dynamically generated targets with
build-cli-plugins-% and build-cli-plugins-%-publish. The first pattern
would match either string. Due to this, jobs that would rely on the
publishing of the plugins would fail since the publish step would never
actually run.

This updates the dynamically created build targets to have distinct name
prefixes. There are now build-cli-plugin-% and publish-cli-plugin-%
targets, so the versions of make on Linux or MacOS work as expected.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>