Add preflight warnings
We had previously only supported returning errors from preflight checks
when certain conditions were not met. We now have experimental support
for ARM64 platforms, so we would like to also have the ability to return
warnings from the results of our preflights.

This plumbs in the ability for our preflight calls to return warning
messages. It also adds a check for arm64 and uses the warning mechanism
to alert the user that this is an experimental platform that may not
entirely work.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>