Handle include links for versioned docs (#3643)
With the switch to versioned docs, our include shortcodes were broken
and that content would not get added to the page. This was due to the
path changing from `/docs/file` to `docs/$version/file`.

This updates the shortcode to automatically adjust the path for the
given version. It also addressed a couple relative references for things
like images.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>