Update docs url version detection for includes
The current implementation of the `include` shortcode was updated to
detect the right path to support versioned docs. There was an error with
this implementation though, with the version always being detected as
the most recent official release. This would cause other versions of the
docs to try to pull in a files from other versions.

This updates the version detection logic to match our sidebar logic to
extract the version from the current path.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>